value: switching to `int32_t' into Integer
The `Integer' class attribute value is change from `long' to `int32_t'
to from the utils/Types.hpp.

- Updating the source code of VLE to remove reference to the `long' into
  the Integer class (remove the toLong, addLong functions.).
- Updating the utils::Tools files to cleanup the conversion function.
  Now we provide explicit template functions:
  - bool is < T >(const std::string& str);
  - T to(const std::string& str);
  - std::string to(T t);
  The template type T can be double, float, bool, int8_t, int16_t,
  int32_t, uint8_t, uint16_t or uint32_t.

+++ src/vle/value/Integer.cpp
@@ -36,7 +36,7 @@ void Integer::writeFile(std::ostream& out) const
 {
     std::streamsize old = out.precision();
 
-    out << std::setprecision(std::numeric_limits < long >::digits10)
+    out << std::setprecision(std::numeric_limits < int32_t >::digits10)
         << m_value;
 
     out.precision(old);
@@ -46,7 +46,7 @@ void Integer::writeString(std::ostream& out) const
 {
     std::streamsize old = out.precision();
 
-    out << std::setprecision(std::numeric_limits < long >::digits10)
+    out << std::setprecision(std::numeric_limits < int32_t >::digits10)
         << m_value;
 
     out.precision(old);
@@ -57,7 +57,7 @@ void Integer::writeXml(std::ostream& out) const
     std::streamsize old = out.precision();
 
     out << "<integer>"
-        << std::setprecision(std::numeric_limits < long >::digits10)
+        << std::setprecision(std::numeric_limits < int32_t >::digits10)
         << m_value
         << "</integer>";
